avoiding reassignment of nans when byteswapping
- mseed: see #485

+++ obspy/mseed/src/libmseed/unpack.c
@@ -840,7 +840,7 @@ msr_unpack_data ( MSRecord *msr, int swapflag, int verbose )
       if ( verbose > 1 )
 	ms_log (1, "%s: Unpacking FLOAT-32 data samples\n", UNPACK_SRCNAME);
       
-      nsamples = msr_unpack_float_32 ((float *)dbuf, (int)msr->samplecnt,
+      nsamples = msr_unpack_float_32 ((uint32_t *)dbuf, (int)msr->samplecnt,
 				      (int)msr->samplecnt, msr->datasamples,
 				      swapflag);
       msr->sampletype = 'f';
@@ -850,7 +850,7 @@ msr_unpack_data ( MSRecord *msr, int swapflag, int verbose )
       if ( verbose > 1 )
 	ms_log (1, "%s: Unpacking FLOAT-64 data samples\n", UNPACK_SRCNAME);
       
-      nsamples = msr_unpack_float_64 ((double *)dbuf, (int)msr->samplecnt,
+      nsamples = msr_unpack_float_64 ((uint64_t *)dbuf, (int)msr->samplecnt,
 				      (int)msr->samplecnt, msr->datasamples,
 				      swapflag);
       msr->sampletype = 'd';
